+++ nodes/encoder.js
@@ -301,10 +301,11 @@ module.exports = {
     * @param {date}   to        Card validity end date
     * @param {object} doors     Object mapping door numbers 1..4 to access permission. A permission
     *                           may be true, false or a time profile in the range [2..254]
+    * @param {number} PIN       Optional card PIN code (in the range 0..999999)
     *
     * @return {buffer} 64 byte NodeJS buffer with encoded put-card request.
     */
-  PutCard: function (deviceId, { card, from, to, doors } = {}) {
+  PutCard: function (deviceId, { card, from, to, doors, PIN } = {}) {
     const request = Buffer.alloc(64)
 
     request.writeUInt8(0x17, 0)
@@ -330,6 +331,18 @@ module.exports = {
         }
       }
     })
+
+    if (PIN) {
+      const pin = Number.parseInt(`${PIN}`)
+
+      if (!Number.isNaN(pin) && pin >= 0 && pin < 1000000) {
+        request.writeUInt8((pin >> 0) & 0x00ff, 24)
+        request.writeUInt8((pin >> 8) & 0x00ff, 25)
+        request.writeUInt8((pin >> 16) & 0x00ff, 26)
+      } else {
+        throw new Error(`invalid card PIN value ${PIN}`)
+      }
+    }
 
     return request
   },